Fix HTML5 Boilerplate link
Ugh, so linking to a line in H5B's master branch doesn't really work,
because the reference could easily be off at any point. So, I picked the
most recent commit as of writing, and linked to the correct line. So now
that link will always be correct, even though it gets stale. I'll take
stale over wrong any day.